requirements.txt
bcrypt>=4.3.0
cffi>=1.17.1
cryptography>=45.0.5
numpy>=2.3.1
paramiko>=3.5.1
pycparser>=2.22
PyNaCl>=1.5.0
PyOpenGL>=3.1.9
PyQt6>=6.9.1
PyQt6-Qt6>=6.9.1
PyQt6_sip>=13.10.2
pyte>=0.8.2
wcwidth>=0.2.13
